Cleaned up swarmStorage
