+++ src/swarm-social/swarmStorage.ts
@@ -2,14 +2,6 @@ import {
     PostCommand,
     PostCommandLog,
     PostCommandLogStorage,
-    PostCommandProtocolVersion,
-    getPreviousCommandEpochFromLog,
-    getHighestSeenTimestampFromLog,
-    getParentUpdateTimestampFromLog,
-    getUnsyncedPostCommandLog,
-    getLatestPostCommandEpochFromLog,
-    epochCompare,
-    mergePostCommandLogs,
 } from '../social/api';
 import { serialize, deserialize } from '../social/serialization';
 import * as Swarm from '../swarm/Swarm';
@@ -19,7 +11,6 @@ import { Post, Author } from '../models/Post';
 import { ImageData } from '../models/ImageData';
 import { ModelHelper } from '../models/ModelHelper';
 import { MockModelHelper } from '../models/__mocks__/ModelHelper';
-import { syncPostCommandLogWithStorage } from '../social/sync';
 
 interface PostOptions {
     shareFeedAddress: boolean;
@@ -208,180 +199,4 @@ export const uploadPosts = async (
         uploadedPosts.push(uploadedPost);
     }
     return uploadedPosts;
-};
-
-const shareNewPostSwarm = async (
-    post: Post,
-    source: string,
-    postCommandLog: PostCommandLog,
-    swarm: Swarm.WriteableApi,
-    options: PostOptions = defaultPostOptions,
-): Promise<PostCommandLog> => {
-    const uploadedPost = await uploadPost(swarm.bzz, post, options.imageResizer, options.modelHelper);
-    const previousEpoch = getPreviousCommandEpochFromLog(postCommandLog);
-    const timestamp = getHighestSeenTimestampFromLog(postCommandLog) + 1;
-    const postCommand: PostCommand = {
-        protocolVersion: PostCommandProtocolVersion,
-        timestamp,
-        parentTimestamp: 0,
-        post: uploadedPost,
-        type: 'update',
-        source,
-        previousEpoch,
-        epoch: undefined,
-    };
-    const uploadedPostCommand =  await addPostCommandToFeed(postCommand, swarm.feed);
-    return {
-        ...postCommandLog,
-        commands: [uploadedPostCommand, ...postCommandLog.commands],
-    };
-};
-
-const updatePostSwarm = async (
-    post: Post,
-    source: string,
-    postCommandLog: PostCommandLog,
-    swarm: Swarm.WriteableApi,
-    options: PostOptions = defaultPostOptions,
-): Promise<PostCommandLog> => {
-    const parentTimestamp = getParentUpdateTimestampFromLog(post, postCommandLog);
-    if (parentTimestamp === 0) {
-        throw new Error('updatePost failed, no previous post with the same id: ' + post._id);
-    }
-    // This is a hack now to force upload to Swarm
-    const updatedPost = {
-        ...post,
-        link: undefined,
-    };
-    const uploadedPost = await uploadPost(swarm.bzz, updatedPost, options.imageResizer, options.modelHelper);
-    const previousEpoch = getPreviousCommandEpochFromLog(postCommandLog);
-    const timestamp = getHighestSeenTimestampFromLog(postCommandLog) + 1;
-    const postCommand: PostCommand = {
-        protocolVersion: PostCommandProtocolVersion,
-        timestamp,
-        parentTimestamp,
-        post: uploadedPost,
-        type: 'update',
-        source,
-        previousEpoch,
-        epoch: undefined,
-    };
-    const uploadedPostCommand =  await addPostCommandToFeed(postCommand, swarm.feed);
-    return {
-        ...postCommandLog,
-        commands: [uploadedPostCommand, ...postCommandLog.commands],
-    };
-};
-
-const removePostSwarm = async (
-    post: Post,
-    source: string,
-    postCommandLog: PostCommandLog,
-    swarmFeedApi: Swarm.WriteableFeedApi,
-) => {
-    const parentTimestamp = getParentUpdateTimestampFromLog(post, postCommandLog);
-    if (parentTimestamp === 0) {
-        throw new Error('removePost failed, no previous post with the same id: ' + post._id);
-    }
-    const timestamp = getHighestSeenTimestampFromLog(postCommandLog) + 1;
-    const previousEpoch = getPreviousCommandEpochFromLog(postCommandLog);
-
-    const removedPost: Post = {
-        _id: post._id,
-        text: '',
-        images: [],
-        createdAt: post.createdAt,
-    };
-    const postCommand: PostCommand = {
-        protocolVersion: PostCommandProtocolVersion,
-        post: removedPost,
-        type: 'remove',
-        source,
-        timestamp,
-        parentTimestamp,
-        previousEpoch,
-    };
-    const removedPostCommand =  await addPostCommandToFeed(postCommand, swarmFeedApi);
-    return {
-        ...postCommandLog,
-        commands: [removedPostCommand, ...postCommandLog.commands],
-    };
-};
-
-const uploadUnsyncedPostCommandsToSwarm = async (postCommandLog: PostCommandLog, swarmApi: Swarm.Api): Promise<PostCommandLog> => {
-    const unsyncedCommandLog = getUnsyncedPostCommandLog(postCommandLog);
-    const syncedCommands = postCommandLog.commands.slice(unsyncedCommandLog.commands.length);
-    const reversedUnsyncedCommands = [...unsyncedCommandLog.commands].reverse();
-
-    const previousSyncedEpoch = getLatestPostCommandEpochFromLog({commands: syncedCommands});
-
-    let previousEpoch = previousSyncedEpoch;
-    const uploadedCommands: PostCommand[] = [];
-    for (const postCommand of reversedUnsyncedCommands) {
-        const postCommandWithPreviousEpoch = {
-            ...postCommand,
-            previousEpoch,
-        };
-
-        const uploadedCommand = await uploadPostCommandToSwarm(postCommandWithPreviousEpoch, swarmApi);
-
-        uploadedCommands.push(uploadedCommand);
-
-        previousEpoch = uploadedCommand.epoch;
-    }
-
-    return {
-        commands: uploadedCommands.reverse().concat(syncedCommands),
-    };
-};
-
-const syncPostCommandLogWithSwarm = async (postCommandLog: PostCommandLog, swarmApi: Swarm.Api): Promise<PostCommandLog> => {
-    const latestEpoch = getLatestPostCommandEpochFromLog(postCommandLog);
-
-    const swarmPostCommandLog = await fetchSwarmPostCommandLog(swarmApi.feed);
-    const swarmLatestEpoch = getLatestPostCommandEpochFromLog(swarmPostCommandLog);
-
-    Debug.log('syncPostCommandLogWithSwarm', latestEpoch, swarmLatestEpoch);
-    if (epochCompare(latestEpoch, swarmLatestEpoch) === 0) {
-        return postCommandLog;
-    }
-
-    const mergedPostCommandLog = mergePostCommandLogs(postCommandLog, swarmPostCommandLog);
-    Debug.log('syncPostCommandLogWithSwarm', 'mergedPostCommandLog', mergedPostCommandLog);
-
-    const uploadedPostCommandLog = await uploadUnsyncedPostCommandsToSwarm(mergedPostCommandLog, swarmApi);
-
-    return uploadedPostCommandLog;
-};
-
-const emptyPostCommandFeed: PostCommandLog = {
-    commands: [],
-};
-
-const testIdentity = {
-    privateKey: '0x12ce6e8759025973fe69dde3873fc2d9e040d79072135ab168369c57589413bc',
-    publicKey: '0x042a8300b3447ffcc27ab32a0e0cf74a8a72022ca51420d9c15ea475d26da40d6ca866ad4504b1943edc784ee96e1f11e84ba62cedcd75ab719dcc626902877a28',
-    address: '0x8f24f61d21a6e3087a7f0b3e158a639a624036cf',
-};
-
-export const defaultSwarmApi = Swarm.makeApi(
-    Swarm.makeFeedAddressFromPublicIdentity(testIdentity),
-    (digest) => Swarm.signDigest(digest, testIdentity),
-);
-
-export const defaultSwarmFeedApi = defaultSwarmApi.feed;
-
-const testSharePostSwarm = async (
-    id: number = 1,
-    postCommandLog: PostCommandLog = emptyPostCommandFeed,
-    source: string = '',
-    swarmApi = defaultSwarmApi,
-): Promise<PostCommandLog> => {
-    const post: Post = {
-        _id: id,
-        text: 'hello' + id,
-        images: [],
-        createdAt: Date.now(),
-    };
-    return await shareNewPostSwarm(post, source, postCommandLog, swarmApi);
 };
